FIX: Specifying verbosity now has the expected outcome
The problem was this: calling a vprinter class object to print something with a specified verbosity would not print that - if the class was instantiated with that exact same verbosity. Example, calling vprinter object with verbosity 2 would not print it, if self.verbosity is itself 2, which is counterintuitive. If the class has been setup with verbosity 2, it should print everything with verbosity less than or equal to 2.